pyproject.toml
[tool.poetry]
name = "kyd_ai-service-reference-implementation"
version = "0.0.1"
description = "Simple reference inplementation of AI service specilized for KYD"
authors = ["VAMStack"]

[tool.poetry.dependencies]
python = "^3.10"
structlog = "^21.5.0"
rich = "^12.4.4"
Flask = "^2.1.2"
requests = "^2.28.1"
python-magic = "^0.4.27"
opencv-python-headless = "^4.6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
